Untar source pkgs into unique dir, and unlink
Source packages were previously extracted directly into the tempdir(),
which could be problematic if directory names collide. Also unlink was
called non-recursively.

These problems are exacerbated if you try to install source packages
where the contents of the package are directly in the root, instead of
having a folder like $packagename wherein the package contents live.
This isn't common but is valid and breaks devtools if you try to install
>1 of these during the same R session.