Ignore code files included in .Rbuildignore
I have a file in my R/ directory that creates the files needed to go in the data/ directory.  I want to keep it with the package, but I don't want it in the package, so I add it to .Rbuildignore.  It takes a long time to run, and I also don't want it source()'d by load_all().  This change makes code files listed in .Rbuildignore be ignored by find_code(), and hence by load_all().