Add uses_github(), extend github_info()
The new unexported function uses_github() is inspired by uses_git() and
provides an explicit test for a package's GitHub-ness. The previous
implicit tests just required a remote named "origin". The new function
checks if there is at least one remote with a URL that contains "github".

github_info() is now more open-minded about the name of the github
remote. The default is "origin", which was previously the only
accepted name. But the names "github" and "upstream" are now
also valid. There's even a new argument to specify the name,
although I'm not sure that's even necessary.