Support passthrough of credentials to git2r
git2r has very sensible defaults for credentials, eg. for ssh it
defaults to looking for ssh-agent, but there are times where it is
useful to specify specific credentials.

This modification simply allows for a git2r credentials object to passet
through to the underlying clone call. This is useful if you wish to use
devtools non-interactively (e.g. a script in a CI solution)