Fixes to ensure package2remote respects an updated .libPath()
packageDescription() loads the description from the namespace if it is
already loaded, which is not what we want for e.g. revdep_check().

This forces the lookup to be on the current .libPaths() only.

The ceiling parameter in gitr2::discover_repository() is used so that
use_git() only returns true if the input path is the base of the git
repository. This prevents test packages from being considered to be
using git.