Do not explicitly set type to 'binary'
Doing so fails if a CRAN-like repository (such as a drat repository) is
used, as they do not have a binary package index.

We preserve the behavior of not prompting and installing the binary even
if there is a newer source by setting
"install.packages.compile.from.source" to "never", which always installs
the binary even if there is a newer source.

https://github.com/wch/r-source/blob/2f2e4711ad7089f97f22c6b1ae25ba582d2e99a6/src/library/utils/R/packages2.R#L408-L436
has details of how that option is used in `install.packages()`.