Do not filter dependencies by those listed in DESCRIPTION
The idea of filtering was to prevent listing a dependency in Remotes:
without also listing it somewhere else in your DESCRIPTION. The
filtering was previously only done on remote dependencies,
7f54aaf4c9f2b31d498d096d98004d4a906b8fb5 inadvertently changed the
filtering for all dependencies which caused the bug.

Thinking about this some more I think we should just remove this
filtering altogether. There can be times you want to install something
in Remotes without actually taking a dependency on it. (covr or lintr
for instance)