Match depends correctly
This change ensures that only exact matches in imports will get ignored. Currently adding e.g. `plyr` to a package that already imports `dplyr` will silently fail. This will happen for all packages that are substrings of packages already ignored.

The proposed change adds regex "fences" around the name of the package to import to circumvent this...