Use trim_ws for all R versions
`onload_assign()` has been moved to pkgload and I don't want to
duplicate it here, so we will just unconditionally use the trimws
implementation from R-3.2. There is a possibility the R-core version
changes, but as it is a simple function any potential changes would
likely not affect the usage in devtools.