Track coverage more robustly
With recent changes to how R dispatches for S3, namely registered
methods are preferred over lexical dispatch we need to track the
coverage on the environment rather than sourcing the new file.

This requires using some un-exported functions in covr. I will hopefully
work on a better interface to avoid this in the future.